More analysis improvements
