+++ scripts/mcmc_run.py
@@ -239,7 +239,6 @@ def main(args):
         if (i == max_samples) or (i % args.output_interval == 0):
             mcmc.analyze(
                 root_dir.parent,
-                os.path.basename(args.config_file),
                 args.datasets,
                 plot_corner=corner,
                 plot_bands=True,
