+++ scripts/pem_mcmc/analysis.py
@@ -6,11 +6,13 @@ import time
 from pathlib import Path
 from typing import Optional
 
+import amisc.distribution as distributions
+import amisc.transform as transform
 import matplotlib.pyplot as plt
 import numpy as np
-from amisc import System
+import scipy.stats
+from amisc import System, Variable
 from matplotlib.axes import Axes
-from matplotlib.figure import Figure
 from scipy.spatial.distance import cdist, euclidean
 
 import hallmd.data
@@ -109,9 +111,40 @@ AXIS_LIMITS = {
 
 data_kwargs = {'markersize': 3.5}
 
+# Sort order for pemv1 variables.
+# For other systems, we do not try and sort the variables
+sort_order = [
+    "T_e",
+    "V_vac",
+    "P_T",
+    "Pstar",
+    "anom_scale",
+    "anom_barrier_scale",
+    "anom_center",
+    "anom_width",
+    "anom_shift_length",
+    "u_n",
+    "c_w",
+    "f_n",
+    "c0",
+    "c1",
+    "c2",
+    "c3",
+    "c4",
+    "c5",
+]
 
 # Quantiles for computing error bars
 QUANTILES = [0.05, 0.25, 0.5, 0.75, 0.95]
+
+
+def try_sort_variables(vars: dict[Variable, np.ndarray], sort_order: list[Variable]) -> dict[Variable, np.ndarray]:
+    # Check that the keys are the same
+    input_vars = set(vars.keys())
+    if input_vars != set(sort_order):
+        return vars
+
+    return {var: vars[var] for var in sort_order}
 
 
 def pad_limits(lims: tuple[float, float], pad: float) -> tuple[float, float]:
@@ -131,8 +164,7 @@ def _stop_timer(start_time: float):
 
 
 def analyze(
-    path,
-    config,
+    amisc_dir,
     datasets,
     plot_map=False,
     plot_corner=False,
@@ -144,30 +176,34 @@ def analyze(
     burn_fraction=0.0,
     limits="posterior",
 ):
-    mcmc_path = Path(path) / "mcmc"
+    dir = Path(amisc_dir)
+    mcmc_path = dir / "mcmc"
     logfile = mcmc_path / "mcmc.csv"
-    plot_path = Path(path) / "mcmc_analysis"
+    plot_path = dir / "mcmc_analysis"
     os.makedirs(plot_path, exist_ok=True)
     print("Generating plots in", plot_path)
 
     analysis_start = time.time()
 
-    system = System.load_from_file(Path(path) / config)
+    system = io.load_system(dir)
     device_name = system['Thruster'].model_kwargs['thruster']
     device = hallmd.utils.load_device(device_name)
 
     variables, samples, logposts, accepted, ids = io.read_output_file(logfile)
-
     num_burn = math.floor(burn_fraction * len(samples))
 
-    samples_raw = np.array(samples)
-    samples = np.array(samples)[num_burn:]
-    logposts = np.array(logposts)
+    samples_raw = samples
+    samples = samples[num_burn:]
     map_ind = np.argmax(logposts)
-    accepted = np.array(accepted)[num_burn:]
+    accepted = accepted[num_burn:]
 
     num_accept = len(samples)
     tex_names = [system.inputs()[name].tex for name in variables]
+
+    sorted_vars = [system.inputs()[var] for var in sort_order]
+    variable_dict = try_sort_variables(
+        {system.inputs()[name]: samples[:, i] for (i, name) in enumerate(variables)}, sorted_vars
+    )
 
     if num_accept > 10:
         start = _start_timer("Computing map, mean, median, and covariance")
@@ -197,36 +233,49 @@ def analyze(
             _stop_timer(start)
 
         if plot_corner:
+            start = _start_timer("Plotting corner plot")
+            # the corner plot sometimes fails early on if there aren't enough distinct samples
             try:
-                start = _start_timer("Plotting corner plot")
-                _plot_corner(samples, tex_names, plot_path, map=map, mean=mean, median=median)
-                _stop_timer(start)
+                make_componentwise_cornerplots(variable_dict, system, plot_path)
             except ValueError:
-                # the corner plot sometimes fails early on if there aren't enough distinct samples
                 pass
+            _stop_timer(start)
 
-    start = _start_timer("Loading data")
-    data = hallmd.data.load(hallmd.data.get_thruster(device_name).datasets_from_names(datasets))
-    channel_length = device['geometry']['channel_length']
-    map = _load_sim_results([ids[map_ind]], mcmc_path)
-    if map:
-        map = map[0]['output']
-    else:
-        map = None
+        # Compute parameter statistics and dump to a LaTeX table
+        compute_sample_statistics(variable_dict, plot_path)
 
-    _stop_timer(start)
-
-    if subsample is not None and len(samples) > subsample:
-        print(f"Subsampling {subsample} samples.")
-        sample_inds = np.random.randint(low=num_burn, high=len(ids) - 1, size=subsample, dtype=int)
-    else:
-        sample_inds = np.arange(num_burn, len(ids))
-
-    # Plot bands
     if plot_bands or calc_metrics:
+        start = _start_timer("Loading data")
+        data = hallmd.data.load(hallmd.data.get_thruster(device_name).datasets_from_names(datasets))
+        channel_length = device['geometry']['channel_length']
+        map = _load_sim_results([ids[map_ind]], mcmc_path)
+
+        if map:
+            map = map[0]['output']
+        else:
+            map = None
+
+        if device_name.casefold() and map is not None == "h9":
+            assert isinstance(map, dict)
+            map = _merge_opconds([map], data)[0]
+
+        _stop_timer(start)
+
+        if subsample is not None and len(samples) > subsample:
+            print(f"Subsampling {subsample} samples.")
+            sample_inds = np.random.randint(low=num_burn, high=len(ids) - 1, size=subsample, dtype=int)
+        else:
+            sample_inds = np.arange(num_burn, len(ids))
+
+        # Plot bands
         start = _start_timer("Loading results")
         results_all = _load_sim_results(np.array(ids)[sample_inds], mcmc_path)
+
         outputs = [res['output'] for res in results_all]
+
+        if device_name.casefold() == "h9":
+            outputs = _merge_opconds(outputs, data)
+
         _stop_timer(start)
 
         _map = map if plot_map else None
@@ -304,7 +353,7 @@ def analyze(
             _stop_timer(start)
 
             start = _start_timer("Plotting ion velocity")
-            uion_median = _plot_ion_vel(
+            uion_median, uion_pressures = _plot_ion_vel(
                 data,
                 outputs,
                 plot_path,
@@ -315,6 +364,11 @@ def analyze(
                 yscalefactor=1 / 1000,
                 thruster=device_name,
             )
+            _stop_timer(start)
+
+            start = _start_timer("Plotting anom. transport")
+            plot_anom(variable_dict, uion_pressures, plot_path)
+
             _stop_timer(start)
 
             start = _start_timer("Plotting ion current density")
@@ -570,7 +624,7 @@ def _plot_ion_cur(
                     handles_data.append((h_sim, h_data))
                     labels_data.append(f"{pressure_str}")
                 else:
-                    h_model = _plot_median_and_uncertainty(axis, angles_sim, qt)
+                    h_model = plot_median_and_uncertainty(axis, angles_sim, qt)
                     ax_solo.legend([h_data, h_model], ["Data", "Model (median + 90\\% CI)"], loc='upper right')
 
             save_figure(fig_solo, subdir, plot_name)
@@ -591,7 +645,7 @@ def _plot_ion_vel(
     yscalefactor: float,
     thruster: str,
     map: Dataset | None = None,
-):
+) -> tuple[dict[hallmd.data.OperatingCondition, hallmd.data.IonVelocityData], list[float]]:
     qty_name = "ion_velocity"
     out_path = plot_path / qty_name
     os.makedirs(out_path, exist_ok=True)
@@ -601,11 +655,11 @@ def _plot_ion_vel(
     mask = np.array([getattr(x, qty_name) is not None for x in data.values()])
     opconds = [opcond for (i, opcond) in enumerate(data.keys()) if mask[i]]
 
-    if len(opconds) == 0:
-        return
-
     # Extract simulation coords and data
     medians = {}
+
+    if len(opconds) == 0:
+        return {}, []
 
     # Individual plots for each pressure
     for i, opcond in enumerate(data.keys()):
@@ -654,7 +708,7 @@ def _plot_ion_vel(
             labels.append("Data")
 
         # Plot median and 90% credible intervals
-        handles.append(_plot_median_and_uncertainty(ax, x_sim, qt))
+        handles.append(plot_median_and_uncertainty(ax, x_sim, qt))
 
         labels += ["Model (median and 90\\% CI)"]
 
@@ -702,6 +756,8 @@ def _plot_ion_vel(
         inds = [0, middle_index, len(opconds) - 1]
     else:
         inds = list(range(len(opconds)))
+
+    pressures = [opconds[ind].background_pressure_torr for ind in inds]
 
     ymin = np.inf
     ymax = -np.inf
@@ -773,7 +829,7 @@ def _plot_ion_vel(
 
     ax.legend(handles, labels, loc=legend_loc)
     save_figure(fig, out_path, plot_name)
-    return medians
+    return medians, pressures
 
 
 def _plot_prediction_accuracy(
@@ -917,7 +973,7 @@ def _plot_global_quantity(
 
         # Sort quantiles by pressure for plotting
         qt = [q[sortperm_sim] for q in qt]
-        handles.append(_plot_median_and_uncertainty(ax, pressure_sim, qt))
+        handles.append(plot_median_and_uncertainty(ax, pressure_sim, qt))
         labels.append("Model")
 
         if map is not None:
@@ -954,9 +1010,28 @@ def _load_sim_results(ids, mcmc_path: Path) -> list[dict]:
     return data
 
 
-def _plot_median_and_uncertainty(ax, x, qt):
-    h_uncertainty = ax.fill_between(x, qt[0], qt[-1], facecolor=COLORS["blue"], zorder=0, alpha=0.25)
-    h_median = ax.plot(x, qt[2], color=COLORS["darkblue"])
+def _merge_opconds(sim: list[dict], data: Dataset) -> list[dict]:
+    """
+    We changed the input mass flow rate for the H9 to fix a bug, but that means the sim mass flow rates no longer
+    match up with the data. This function fixes that by finding the conditions with the same pressures as the data.
+    """
+
+    sim_to_data = {}
+
+    for sim_opcond in sim[0].keys():
+        for data_opcond in data.keys():
+            if sim_opcond.background_pressure_torr == data_opcond.background_pressure_torr:
+                sim_to_data[sim_opcond] = data_opcond
+                break
+
+    return [{sim_to_data[k]: v for (k, v) in s.items()} for s in sim]
+
+
+def plot_median_and_uncertainty(
+    ax, x, qt, alpha=0.25, lightcolor=COLORS["blue"], darkcolor=COLORS["darkblue"], linestyle='-'
+):
+    h_uncertainty = ax.fill_between(x, qt[0], qt[-1], facecolor=lightcolor, zorder=0, alpha=alpha)
+    h_median = ax.plot(x, qt[2], color=darkcolor, linestyle=linestyle)
     return h_uncertainty, h_median[0]
 
 
@@ -1002,100 +1077,276 @@ def _plot_traces(names, samples, dir: Path = Path(".")):
     return
 
 
-def _plot_corner(
-    samples: np.ndarray,
-    var_names: list[str],
-    dir: Path = Path("."),
-    map: np.ndarray | None = None,
-    mean: np.ndarray | None = None,
-    median: np.ndarray | None = None,
-) -> tuple[Figure, Axes]:
-    fontsize = 12
-    num_vars = len(var_names)
-    size = 3 * num_vars
-    fig, axes = plt.subplots(num_vars, num_vars, figsize=(size, size))
+def make_componentwise_cornerplots(samples: dict[Variable, np.ndarray], system: System, output_dir: Path):
+    component_vars = split_vars_by_components(samples, system)
 
-    lims = [_determine_limits(samples[:, i]) for i in range(num_vars)]
+    # Plot individual corner plots per-component
+    for comp_name, var_dict in component_vars.items():
+        make_cornerplot(var_dict, output_dir, comp_name.casefold())
 
-    map_color = "red"
-    mean_color = "orange"
-    median_color = "pink"
 
-    for j, row in enumerate(axes):
-        for i, ax in enumerate(row):
-            if i > j:
-                ax.set_axis_off()
+def make_cornerplot(samples: dict[Variable, np.ndarray], output_dir: Path, plot_name: str = ""):
+    num_vars = len(samples.keys())
+    subfig_size = 2.0
+    rows = num_vars
+
+    fig, axes = plt.subplots(
+        rows,
+        rows,
+        dpi=200,
+        figsize=(rows * subfig_size, rows * subfig_size),
+        layout="constrained",
+    )
+
+    limits = [pad_limits((np.min(samples), np.max(samples)), 0.2) for samples in samples.values()]
+
+    for irow, rowvar in enumerate(samples.keys()):
+        for icol, colvar in enumerate(samples.keys()):
+            axis = axes[irow, icol]
+            if irow < icol:
+                axis.set_visible(False)
                 continue
-            elif i == j:
-                _ax_hist1d(ax, samples[:, i], lims[i])
-                ax.set_yticks([])
 
-                if mean is not None:
-                    ax.axvline(mean[i], color=mean_color, linewidth=2)
+            axis.grid(False)
 
-                if median is not None:
-                    ax.axvline(median[i], color=median_color, linewidth=2)
+            if irow == rows - 1:
+                axis.set_xlabel(colvar.tex)
+            else:
+                axis.set_xticklabels([])
 
-                if map is not None:
-                    ax.axvline(map[i], color=map_color, linewidth=2)
+            if icol == 0 and irow > 0:
+                axis.set_ylabel(rowvar.tex)
+            else:
+                axis.set_yticklabels([])
 
-                if j == num_vars - 1:
-                    ax.set_xlabel(var_names[i], fontsize=fontsize)
-                else:
-                    ax.set_xticklabels([])
+            axis.set_xlim(limits[icol])
 
-            elif i < j:
-                x = samples[:, i]
-                y = samples[:, j]
-                nbins_x = _num_bins(x, lims[i])
-                nbins_y = _num_bins(y, lims[j])
-                ax.hist2d(samples[:, i], samples[:, j], bins=[nbins_x, nbins_y], range=[lims[i], lims[j]])
+            if irow == icol:
+                # 1D histogram - kernel density estimate
+                _axis_hist_1D(axis, samples[colvar], limits[icol])
+                axis.set_yticks([])
+                axis.spines['top'].set_visible(False)
+                axis.spines['right'].set_visible(False)
+                axis.spines['left'].set_visible(False)
+                axis.set_ylim(bottom=0)
+                continue
+            else:
+                axis.hexbin(samples[colvar], samples[rowvar], gridsize=12, mincnt=1)
+                axis.set_ylim(limits[irow])
 
-                if mean is not None:
-                    ax.scatter([mean[i]], [mean[j]], color=mean_color)
+    if plot_name:
+        plot_name = "corner_" + plot_name
+    else:
+        plot_name = "corner"
 
-                if median is not None:
-                    ax.scatter([median[i]], [median[j]], color=median_color)
-
-                if map is not None:
-                    ax.scatter([map[i]], [map[j]], color=map_color)
-
-                if i == 0:
-                    ax.set_ylabel(var_names[j], fontsize=fontsize)
-                else:
-                    ax.set_yticklabels([])
-
-                if j == num_vars - 1:
-                    ax.set_xlabel(var_names[i], fontsize=fontsize)
-                else:
-                    ax.set_xticklabels([])
-
-            ax.tick_params(axis="x", rotation=-45)
-
-    plt.tight_layout()
-    fig.savefig(dir / "corner.png", dpi=100)
-    plt.close(fig)
-    return fig, axes
+    save_figure(fig, output_dir, plot_name, tight_layout=False)
 
 
-def _num_bins(samples: np.ndarray, lims: tuple[float, float]) -> int:
-    """Calculate optimal histogram bin count using Friedman-Draconis rule"""
-    q3 = np.percentile(samples, 75)
-    q1 = np.percentile(samples, 25)
-    iqr = q3 - q1
-
-    bin_width = 2 * iqr / np.cbrt(samples.size)
-    num_bins = np.ceil((lims[1] - lims[0]) / bin_width).astype(int)
-    return num_bins
+def _axis_hist_1D(axis: Axes, samples: np.ndarray, lims: tuple[float, float], N: int = 100):
+    kernel = scipy.stats.gaussian_kde(samples)
+    xvals = np.linspace(lims[0], lims[1], N)
+    probs = kernel(xvals)
+    axis.fill_between(xvals, np.zeros(N), probs, color=COLORS["lightblue"])
+    axis.plot(xvals, probs, linewidth=3, color=COLORS["blue"])
+    axis.set_xlim(lims)
 
 
-def _ax_hist1d(ax: Axes, samples: np.ndarray, xlims: tuple[float, float]) -> None:
-    ax.set_xlim(xlims)
-    nbins = _num_bins(samples, xlims)
-    ax.hist(samples, nbins, color="lightgrey", density=True)
+def split_vars_by_components(vars: dict[Variable, np.ndarray], system: System) -> dict[str, dict[Variable, np.ndarray]]:
+    return {
+        comp.name: {
+            v: samples
+            for (v, samples) in vars.items()
+            if v in comp.inputs and not (comp.name == "Thruster" and v.name == "T_e")
+        }
+        for comp in system.components
+    }
 
 
-def _determine_limits(x: np.ndarray) -> tuple[float, float]:
-    min = np.min(x)
-    max = np.max(x)
-    return min, max
+def compute_var_statistics(var: Variable, samples: np.ndarray):
+    quantiles = np.quantile(samples, QUANTILES)
+
+    transform_str = "None" if var.norm is None else f"{var.norm[0]}"
+
+    return {
+        "dist": f"{var.distribution}",
+        "transform": transform_str,
+        "min": np.min(samples),
+        "5%": quantiles[0],
+        "50%": quantiles[2],
+        "95%": quantiles[4],
+        "max": np.max(samples),
+        "std": np.std(samples),
+    }
+
+
+def rounded_number(x):
+    if type(x) is str:
+        if x == '':
+            x = 0
+    f = float(x)
+    if f.is_integer():
+        return int(f), True
+
+    if abs(f - round(f)) < 0.01:
+        return rounded_number(round(f))[0], True
+
+    if abs(f - round(f, ndigits=1)) < 0.001:
+        return round(f, ndigits=1), True
+    return f, False
+
+
+def format_number(x, check_pi=False):
+    f, was_rounded = rounded_number(x)
+    if was_rounded:
+        return f"{f}"
+
+    # check for multiples of pi
+    if check_pi:
+        if f > np.pi:
+            div = f / np.pi
+            if abs(div - round(div)) < 0.01:
+                return f"\\{round(div)}\\pi"
+
+        if f < np.pi:
+            # check for fractions of pi
+            fractions = [2, 3, 4, 6]
+            for frac in fractions:
+                if abs(np.pi / frac - f) < 0.01:
+                    return f"\\pi/{frac}"
+
+    return f"{f:.2f}"
+
+
+def compute_sample_statistics(samples: dict[Variable, np.ndarray], output_dir: Path):
+    output = {var.name: compute_var_statistics(var, _samples) for (var, _samples) in samples.items()}
+
+    with open(output_dir / "variable_stats.json", "w") as fd:
+        json.dump(output, fd)
+
+    cols = 8
+    col_str = " ".join("l" * cols)
+    indent = "    "
+
+    def nth(n):
+        return f"${n}^{{\\text{{th}}}}$ pctile"
+
+    # Create latex table
+    latex_header = f"""\\begin{{tabular}}{{{col_str}}}
+{indent}\\hline
+{indent}\\multicolumn{{2}}{{l}}{{}} & \\multicolumn{{{cols - 2}}}{{l}}{{Posterior}} \\\\ \\cline{{3-{cols}}}
+{indent}Variable & Prior & Min & {nth(5)} & {nth(50)} & {nth(95)} & Max & Std dev \\\\ \\hline"""
+
+    notes = [
+        "Variables with the ($10^x$) notation indicate a log-uniform distribution.",
+        "The ($x$) notation indicates that the variable has been normalized by $x$.",
+    ]
+
+    note_str = "\\\\\n".join([indent + f"\\multicolumn{{{cols}}}{{l}}{{{note}}}" for note in notes])
+
+    latex_footer = f"{indent}\\hline\n{note_str}\n\\end{{tabular}}"
+
+    rows = []
+
+    for var, stats in zip(samples.keys(), output.values()):
+        var_str = f"{var.tex}"
+
+        dist = var.distribution
+
+        if isinstance(dist, distributions.LogUniform):
+            var_str += " ($10^{x}$)"
+
+        if isinstance(dist, distributions.Uniform) or isinstance(dist, distributions.LogUniform):
+            lo, hi = dist.dist_args
+            lo = format_number(var.normalize(lo), check_pi=True)
+            hi = format_number(var.normalize(hi), check_pi=True)
+            dist_str = f"$\\mathcal{{U}}({lo}, {hi})$"
+        else:
+            dist_str = ""
+
+        if var.norm is not None and isinstance(var.norm[0], transform.Linear):
+            exponent = format_number(np.log10(var.norm[0].transform_args[0]))
+            var_str += f" (\\num{{e{exponent}}})"
+
+        min_str = format_number(stats['min'])
+        max_str = format_number(stats['max'])
+        std_str = format_number(stats['std'])
+        pct_5 = format_number(stats['5%'])
+        pct_50 = format_number(stats['50%'])
+        pct_95 = format_number(stats['95%'])
+
+        cols = [var_str, dist_str, min_str, pct_5, pct_50, pct_95, max_str, std_str]
+        row_str = " & ".join(cols) + "\\\\"
+
+        rows.append(indent + row_str)
+
+    with open(output_dir / "variable_table.tex", "w") as fd:
+        print(latex_header, file=fd)
+        print("\n".join(rows), file=fd)
+        print(latex_footer, file=fd)
+
+
+def simple_pressure_shift(pressure, shift_length, slope=2.0, midpoint=25e-6):
+    ratio = pressure / midpoint
+    return -shift_length * (1 / (1 + np.exp(-slope * (ratio - 1))) - 1 / (1 + np.exp(slope)))
+
+
+def scaled_gaussian_bohm(coords, anom_scale, anom_barrier_scale, anom_center, anom_width, z_shift=0.0):
+    return anom_scale * (1 - anom_barrier_scale * np.exp(-0.5 * ((coords - z_shift - anom_center) / (anom_width)) ** 2))
+
+
+def plot_anom(samples: dict[Variable, np.ndarray], pressures: list[float], output_dir: Path):
+    var_name_dict = {v.name: s for (v, s) in samples.items()}
+    anom_scale = var_name_dict["anom_scale"]
+    anom_barrier_scale = var_name_dict["anom_barrier_scale"]
+    anom_center = var_name_dict["anom_center"]
+    anom_width = var_name_dict["anom_width"]
+    shift_length = var_name_dict["anom_shift_length"]
+
+    lightcolors = ["lightred", "lightgreen", "lightblue"]
+    darkcolors = ["red", "green", "darkblue"]
+    linestyles = ['-', '--', '-.']
+
+    if len(pressures) > 3:
+        p = pressures[:3]
+    else:
+        p = pressures
+
+    fig, axis = plt.subplots(figsize=(5, 4.5), dpi=200)
+    left = 0
+    right = 3
+    coords = np.linspace(left, right, 100)
+    axis.set_xlim(left, right)
+    axis.set_xlabel("Axial coordinate [channel lengths]")
+    axis.set_ylabel("$\\nu_{anom} / (\\omega_{ce} / 16)$")
+
+    handles = []
+    labels = []
+
+    for pressure, light, dark, style in zip(p, lightcolors, darkcolors, linestyles):
+        anoms = []
+
+        for i, (alpha, beta, z, L, dz) in enumerate(
+            zip(anom_scale, anom_barrier_scale, anom_center, anom_width, shift_length)
+        ):
+            z_shift = simple_pressure_shift(pressure, dz)
+            anoms.append(scaled_gaussian_bohm(coords, alpha, beta, z, L, z_shift=z_shift) * 16)
+
+        quantiles = np.quantile(np.array(anoms), QUANTILES, axis=0)
+
+        handles.append(
+            plot_median_and_uncertainty(
+                axis,
+                coords,
+                quantiles,
+                alpha=0.5,
+                lightcolor=COLORS[light],
+                darkcolor=COLORS[dark],
+                linestyle=style,
+            )
+        )
+
+        labels.append(f"{pressure / 1e-6:.1f} $\\mu$Torr")
+
+    axis.legend(handles, labels)
+    axis.set_ylim(bottom=0, top=2)
+    save_figure(fig, output_dir, "anom")
