+++ scripts/pem_mcmc/io.py
@@ -1,8 +1,35 @@
+import math
+import os
 from os import PathLike
+from pathlib import Path
 
 import numpy as np
+from amisc import System
 
 from pem_mcmc.types import Array, Number
+
+
+def load_system(directory: Path) -> System:
+    """
+    Given a directory, find the first yaml file in that directory and load the amisc system from that file.
+    """
+    dir_contents = os.listdir(directory)
+    for file_or_dir in dir_contents:
+        # Skip directories
+        if os.path.isdir(file_or_dir):
+            continue
+
+        # Skip files without .yml or .yaml extensions
+        file = Path(file_or_dir)
+        ext = file.suffix.casefold()
+        if ext not in {".yml", ".yaml"}:
+            continue
+
+        # Load the system from the file
+        return System.load_from_file(directory / file)
+
+    # If we're here, we didn't find anything and should error.
+    raise ValueError(f"Could not find a yaml file in directory {directory}.")
 
 
 def append_sample_row(
@@ -28,7 +55,7 @@ def read_dlm(file: PathLike, delimiter: str | None = ',', comments='#') -> dict[
     return {col_name: column for (col_name, column) in zip(col_names, columns)}
 
 
-def read_output_file(path: PathLike):
+def read_output_file(path: PathLike, burn_fraction: float = 0.0):
     """
     Read an output file generated by MCMC.
     TODO: use read_dlm for this.
@@ -81,4 +108,12 @@ def read_output_file(path: PathLike):
             samples.append(np.array([float(x) for x in fields[var_start:var_end]]))
             accepted.append(accept)
 
-    return variables, samples, logposts, accepted, ids
+    # Discard burned samples
+    if burn_fraction > 0:
+        num_burn = math.floor(burn_fraction * len(samples))
+        samples = samples[num_burn:]
+        logposts = logposts[num_burn:]
+        accepted = accepted[num_burn:]
+        ids = ids[num_burn:]
+
+    return variables, np.array(samples), np.array(logposts), np.array(accepted), ids
